rpc: prevent circular deps by extracting into separate include
