+++ src/Makefile.test_fuzz.include
@@ -18,6 +18,7 @@ libtest_fuzz_a_SOURCES = \
   test/fuzz/fuzz.cpp \
   $(TEST_FUZZ_H)
 
+LIBTEST_FUZZ += $(LIBBITCOIN_SERVER_INTERFACES)
 LIBTEST_FUZZ += $(LIBBITCOIN_SERVER)
 LIBTEST_FUZZ += $(LIBBITCOIN_COMMON)
 LIBTEST_FUZZ += $(LIBBITCOIN_UTIL)
