+++ src/Makefile.test_util.include
@@ -30,6 +30,7 @@ libtest_util_a_SOURCES = \
   test/util/wallet.cpp \
   $(TEST_UTIL_H)
 
+LIBTEST_UTIL += $(LIBBITCOIN_SERVER_INTERFACES)
 LIBTEST_UTIL += $(LIBBITCOIN_SERVER)
 LIBTEST_UTIL += $(LIBBITCOIN_COMMON)
 LIBTEST_UTIL += $(LIBBITCOIN_UTIL)
