+++ src/httprpc.cpp
@@ -7,8 +7,8 @@
 #include <chainparams.h>
 #include <crypto/hmac_sha256.h>
 #include <httpserver.h>
+#include <rpc/interfaces.h>
 #include <rpc/protocol.h>
-#include <rpc/server.h>
 #include <ui_interface.h>
 #include <util/strencodings.h>
 #include <util/system.h>
@@ -22,6 +22,8 @@
 #include <stdio.h>
 #include <set>
 #include <string>
+
+extern CRPCTable tableRPC;
 
 #include <boost/algorithm/string.hpp> // boost::trim
