+++ src/rest.cpp
@@ -12,8 +12,8 @@
 #include <primitives/block.h>
 #include <primitives/transaction.h>
 #include <rpc/blockchain.h>
+#include <rpc/interfaces.h>
 #include <rpc/protocol.h>
-#include <rpc/server.h>
 #include <streams.h>
 #include <sync.h>
 #include <txmempool.h>
