+++ src/rpc/net.cpp
@@ -2,7 +2,6 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <rpc/server.h>
 
 #include <banman.h>
 #include <clientversion.h>
@@ -16,6 +15,7 @@
 #include <policy/settings.h>
 #include <rpc/blockchain.h>
 #include <rpc/protocol.h>
+#include <rpc/interfaces.h>
 #include <rpc/util.h>
 #include <sync.h>
 #include <timedata.h>
