+++ src/util/system.cpp
@@ -68,9 +68,6 @@
 #include <typeinfo>
 #include <univalue.h>
 
-// Application startup time (used for uptime calculation)
-const int64_t nStartupTime = GetTime();
-
 const char * const BITCOIN_CONF_FILENAME = "bitcoin.conf";
 
 ArgsManager gArgs;
@@ -1142,12 +1139,6 @@ std::string CopyrightHolders(const std::string& strPrefix)
         strCopyrightHolders += "\n" + strPrefix + "The Bitcoin Core developers";
     }
     return strCopyrightHolders;
-}
-
-// Obtain the application startup time (used for uptime calculation)
-int64_t GetStartupTime()
-{
-    return nStartupTime;
 }
 
 fs::path AbsPathForConfigVal(const fs::path& path, bool net_specific)
