+++ src/wallet/test/wallet_tests.cpp
@@ -12,7 +12,7 @@
 #include <interfaces/chain.h>
 #include <node/context.h>
 #include <policy/policy.h>
-#include <rpc/server.h>
+#include <rpc/interfaces.h>
 #include <test/util/logging.h>
 #include <test/util/setup_common.h>
 #include <util/translation.h>
