+++ src/zmq/zmqrpc.cpp
@@ -4,7 +4,6 @@
 
 #include <zmq/zmqrpc.h>
 
-#include <rpc/server.h>
 #include <rpc/util.h>
 #include <zmq/zmqabstractnotifier.h>
 #include <zmq/zmqnotificationinterface.h>
